rename: api 폴더를 app 폴더 아래에서 src 폴더 아래로 이동
